merges some logic. still need to handle events for RPC
