+++ src/main/java/org/adridadou/ethereum/provider/PrivateEthereumFacadeProvider.java
@@ -2,9 +2,9 @@ package org.adridadou.ethereum.provider;
 
 import com.typesafe.config.ConfigFactory;
 import org.adridadou.ethereum.EthereumFacade;
-import org.adridadou.ethereum.blockchain.EthereumProxyEthereumJ;
+import org.adridadou.ethereum.blockchain.EthereumBackend;
+import org.adridadou.ethereum.blockchain.EthereumProxy;
 import org.adridadou.ethereum.blockchain.EthereumJReal;
-import org.adridadou.ethereum.blockchain.Ethereumj;
 import org.adridadou.ethereum.converters.input.InputTypeHandler;
 import org.adridadou.ethereum.converters.output.OutputTypeHandler;
 import org.adridadou.ethereum.event.EthereumEventHandler;
@@ -58,7 +58,7 @@ public class PrivateEthereumFacadeProvider {
 
         MinerConfig.dbName = config.getDbName();
         Ethereum ethereum = EthereumFactory.createEthereum(MinerConfig.class);
-        Ethereumj ethereumj = new EthereumJReal(ethereum, ChainId.id(1));
+        EthereumBackend ethereumBackend = new EthereumJReal(ethereum, ChainId.id(1));
         ethereum.initSyncing();
 
         if (!dagCached) {
@@ -71,10 +71,10 @@ public class PrivateEthereumFacadeProvider {
             }
         }
 
-        EthereumEventHandler ethereumListener = new EthereumEventHandler(ethereumj);
+        EthereumEventHandler ethereumListener = new EthereumEventHandler(ethereumBackend);
         InputTypeHandler inputTypeHandler = new InputTypeHandler();
         OutputTypeHandler outputTypeHandler = new OutputTypeHandler();
-        final EthereumFacade facade = new EthereumFacade(new EthereumProxyEthereumJ(ethereumj, ethereumListener, inputTypeHandler, outputTypeHandler),inputTypeHandler, outputTypeHandler, SwarmService.from(SwarmService.PUBLIC_HOST), SolidityCompiler.getInstance());
+        final EthereumFacade facade = new EthereumFacade(new EthereumProxy(ethereumBackend, ethereumListener, inputTypeHandler, outputTypeHandler),inputTypeHandler, outputTypeHandler, SwarmService.from(SwarmService.PUBLIC_HOST), SolidityCompiler.getInstance());
 
         //This event does not trigger when you are the miner
         ethereumListener.onSyncDone(EthereumListener.SyncState.COMPLETE);
